pipeline/src/requirements.txt
alembic>=1.2.1
avro-python3>=1.9.1
flask-restplus>=0.13.0
Flask-SQLAlchemy>=2.4.0
Flask-Migrate>=2.5.2
Flask-Script>=2.0.6
jsonify>=0.5
jsonschema>=3.0.2
marshmallow>=3.2.1
numpy>=1.15.4
openpyxl>=3.0.3
pandas>=0.25.1
Pillow>=6.1.0
psycopg2>=2.8.3
python-dateutil>=2.7.5
pytz>=2019.2
regex>=2018.1.10
requests>=2.20.1
rsa>=3.4.2
SQLAlchemy>=1.3.8
tqdm>=4.28.1
urllib3>=1.24.1
Werkzeug==0.15.6
